requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1
djangorestframework==3.15.2
Flask==2.2.5
Flask-Cors==5.0.0
idna==3.8
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.9.0
requests==2.31.0
sqlparse==0.5.1
typing_extensions==4.7.1
urllib3==2.0.7
websockets==13.0.1
Werkzeug==2.2.3
zipp==3.15.0
